test: isolate shared gate fixture state
